Implement form and view to allow forecast by document number

+++ lineitems/models.py
@@ -288,14 +288,14 @@ class LineForecast(models.Model):
             return True
         return False
 
-    def forecast_line_by_line(self, docno: str, forecast: int):
+    def forecast_line_by_line(self, docno: str, forecast: float):
         lines = LineItem.objects.filter(docno=docno)
         document_working_plan = lines.aggregate(models.Sum("workingplan"))["workingplan__sum"]
-        ratio = forecast / document_working_plan
+        ratio = forecast / float(document_working_plan)
         for li in lines:
             if hasattr(li, "fcst"):
                 li_fcst = LineForecastManager().get_line_forecast(li)
-                li_fcst.forecastamount = li.workingplan * ratio
+                li_fcst.forecastamount = float(li.workingplan) * ratio
                 li_fcst.save()
             else:
                 li_fcst = LineForecast(lineitem=li, forecastamount=li.workingplan * ratio)
